Convert valid_version? to a helper method
